+++ NINA.Plugin.Assistant/NINA.Plugin.Assistant/Sequencer/ImageSaveWatcher.cs
@@ -1,12 +1,15 @@
 ﻿using Assistant.NINAPlugin.Database;
 using Assistant.NINAPlugin.Database.Schema;
 using Assistant.NINAPlugin.Plan;
+using Assistant.NINAPlugin.Util;
 using NINA.Plugin.Assistant.Shared.Utility;
+using NINA.Profile;
 using NINA.Profile.Interfaces;
 using NINA.WPF.Base.Interfaces.Mediator;
 using System;
 using System.Collections.Concurrent;
 using System.Data.Entity.Migrations;
+using System.IO;
 using System.Text;
 using System.Threading;
 
@@ -28,7 +31,10 @@ namespace Assistant.NINAPlugin.Sequencer {
     /// </summary>
     public class ImageSaveWatcher : IImageSaveWatcher {
 
+        public static readonly string REJECTED_SUBDIR = "rejected";
+
         private IProfile profile;
+        private ProfilePreference profilePreference;
         private IImageSaveMediator imageSaveMediator;
         private ConcurrentDictionary<int, int> exposureDictionary;
 
@@ -39,6 +45,7 @@ namespace Assistant.NINAPlugin.Sequencer {
 
         public ImageSaveWatcher(IProfile profile, IImageSaveMediator imageSaveMediator, IPlanTarget planTarget, bool synchronizationEnabled) {
             this.profile = profile;
+            this.profilePreference = new SchedulerPlanLoader(profile).GetProfilePreferences();
             this.imageSaveMediator = imageSaveMediator;
             exposureDictionary = new ConcurrentDictionary<int, int>(Environment.ProcessorCount * 2, 31);
             this.planTarget = planTarget;
@@ -87,8 +94,14 @@ namespace Assistant.NINAPlugin.Sequencer {
 
             bool accepted = false;
             string rejectReason = "not graded";
+
             if (enableGrader) {
                 (accepted, rejectReason) = new ImageGrader(profile).GradeImage(planTarget, msg);
+                if (!accepted && profilePreference.EnableMoveRejected) {
+                    string dstDir = Path.Combine(Path.GetDirectoryName(msg.PathToImage.LocalPath), REJECTED_SUBDIR);
+                    TSLogger.Debug($"moving rejected image to {dstDir}");
+                    Utils.MoveFile(msg.PathToImage.LocalPath, dstDir);
+                }
             }
 
             int? imageId = msg.MetaData?.Image?.Id;