+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Target Scheduler
 
 ## 4.0.0.0 - 2023-11-XX
-* Added image grading on FWHM and Eccentricity (required Hocus Focus plugin)
+* Added image grading on FWHM and Eccentricity (requires Hocus Focus plugin)
+* Support for optionally moving rejected images to a 'rejected' directory
 * Added experimental support for synchronization across multiple instances of NINA
 * The rule weight list is now sorted when displayed
 * Added target rotation and ROI to the set of data saved for acquired images.  A future release will use these values when selecting 'like' images for grading.